Rewrote how functions are processed This update changes how functions that run per tick work. Instead of running multiple functions with several commands in them, only a few commands are run to check if those functions should be running. As an example, before the drop_check function which controls how the config menu works would always run, when it would only be used during the config stage of the game. Now the tick function will check to see if config mode is currently active before running
This change significantly optimizes the game and should allow the game
to run much better, especially in areas with a lot of entities.

The crafting detector function has also been set to run every 10 ticks.
Although this adds a slight delay to crafting, it's barely noticeable
and it should give some much better preformance as well.